SchoolManagerApi/requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
coreapi==2.3.3